requirements.txt
crewai>=1.0.0
openai>=1.0.0
google-search-results>=2.4.2
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
schedule>=1.2.0

